Rename package to tests
