+++ graphannis/src/annis/db/aql/conjunction.rs
@@ -1,5 +1,5 @@
 #[cfg(test)]
-mod test;
+mod tests;
 
 use super::disjunction::Disjunction;
 use super::Config;
